Refs #28800 -- Lifted some url functions from admindocs into urls.utils.
