+++ django/contrib/admindocs/views.py
@@ -7,22 +7,16 @@ from django.apps import apps
 from django.contrib import admin
 from django.contrib.admin.views.decorators import staff_member_required
 from django.contrib.admindocs import utils
-from django.contrib.admindocs.utils import (
-    remove_non_capturing_groups,
-    replace_metacharacters,
-    replace_named_groups,
-    replace_unnamed_groups,
-)
 from django.contrib.auth import get_permission_codename
 from django.core.exceptions import (
     ImproperlyConfigured,
     PermissionDenied,
-    ViewDoesNotExist,
 )
 from django.db import models
 from django.http import Http404
 from django.template.engine import Engine
 from django.urls import get_mod_func, get_resolver, get_urlconf
+from django.urls.utils import extract_views_from_urlpatterns, simplify_regex
 from django.utils._os import safe_join
 from django.utils.decorators import method_decorator
 from django.utils.functional import cached_property
@@ -475,49 +469,3 @@ def get_readable_field_data_type(field):
     the values of field.__dict__ before being output.
     """
     return field.description % field.__dict__
-
-
-def extract_views_from_urlpatterns(urlpatterns, base="", namespace=None):
-    """
-    Return a list of views from a list of urlpatterns.
-
-    Each object in the returned list is a 4-tuple:
-    (view_func, regex, namespace, name)
-    """
-    views = []
-    for p in urlpatterns:
-        if hasattr(p, "url_patterns"):
-            try:
-                patterns = p.url_patterns
-            except ImportError:
-                continue
-            views.extend(
-                extract_views_from_urlpatterns(
-                    patterns,
-                    base + str(p.pattern),
-                    (namespace or []) + (p.namespace and [p.namespace] or []),
-                )
-            )
-        elif hasattr(p, "callback"):
-            try:
-                views.append((p.callback, base + str(p.pattern), namespace, p.name))
-            except ViewDoesNotExist:
-                continue
-        else:
-            raise TypeError(_("%s does not appear to be a urlpattern object") % p)
-    return views
-
-
-def simplify_regex(pattern):
-    r"""
-    Clean up urlpattern regexes into something more readable by humans. For
-    example, turn "^(?P<sport_slug>\w+)/athletes/(?P<athlete_slug>\w+)/$"
-    into "/<sport_slug>/athletes/<athlete_slug>/".
-    """
-    pattern = remove_non_capturing_groups(pattern)
-    pattern = replace_named_groups(pattern)
-    pattern = replace_unnamed_groups(pattern)
-    pattern = replace_metacharacters(pattern)
-    if not pattern.startswith("/"):
-        pattern = "/" + pattern
-    return pattern
